feat : feature add
CheckSystem class add
Lotto class modify :
contrastNumber(), tatisticsDisplay()
enum class add : Place 1 ~ 5
